+++ parkhangs-webapp/src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { render } from 'react-dom'
 import { configureStore } from '@reduxjs/toolkit'
 import { Provider } from 'react-redux'
-import App from './components/App'
+import App from './App'
 import rootReducer from './reducers'
 
 const store = configureStore({
